wip: tail logs for CAS tests
