+++ .github/workflows/fit_WA_SRA_CASClient.yml
@@ -43,9 +43,8 @@ jobs:
         restore-keys: ${{ runner.os }}-m2
     - name: Build
       run: mvn -U -T 1C -P 'skipTests,all'
-    - name: Setup tmate session
-      uses: mxschmitt/action-tmate@v3
-    - name: 'Tail logs directory'
-      run: tail -f --retry --lines=50 fit/wa-reference/target/logs &
     - name: 'WA / SRA / CAS Client'
-      run: mvn -f fit/wa-reference/pom.xml verify -Dinvoker.streamLogs=true -Dmodernizer.skip=true -Drat.skip=true -Dcheckstyle.skip=true -Djacoco.skip=true -Dit.test=org.apache.syncope.fit.sra.CASSRAITCase
+      continue-on-error: true
+      run: exit 1
+    - name: Output logs
+      run: echo "This should now run"
