+++ core/starter/src/main/java/org/apache/syncope/core/starter/SyncopeCoreApplication.java
@@ -22,11 +22,14 @@ import java.util.Map;
 import org.apache.cxf.spring.boot.autoconfigure.openapi.OpenApiAutoConfiguration;
 import org.apache.syncope.common.keymaster.client.api.model.NetworkService;
 import org.apache.syncope.common.keymaster.client.api.startstop.KeymasterStop;
+import org.apache.syncope.common.lib.info.SystemInfo;
 import org.apache.syncope.core.persistence.api.DomainHolder;
 import org.apache.syncope.core.starter.actuate.DomainsHealthIndicator;
 import org.apache.syncope.core.starter.actuate.ExternalResourcesHealthIndicator;
+import org.apache.syncope.core.starter.actuate.DefaultSyncopeCoreInfoContributor;
 import org.apache.syncope.core.starter.actuate.SyncopeCoreInfoContributor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.actuate.mail.MailHealthIndicator;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -39,7 +42,9 @@ import org.springframework.boot.autoconfigure.quartz.QuartzAutoConfiguration;
 import org.springframework.boot.autoconfigure.web.servlet.error.ErrorMvcAutoConfiguration;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.context.PayloadApplicationEvent;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.event.EventListener;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
@@ -51,7 +56,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
     DataSourceAutoConfiguration.class,
     DataSourceTransactionManagerAutoConfiguration.class,
     JdbcTemplateAutoConfiguration.class,
-    QuartzAutoConfiguration.class })
+    QuartzAutoConfiguration.class }, proxyBeanMethods = false)
 @EnableTransactionManagement
 public class SyncopeCoreApplication extends SpringBootServletInitializer {
 
@@ -72,7 +77,7 @@ public class SyncopeCoreApplication extends SpringBootServletInitializer {
     @ConditionalOnMissingBean
     @Bean
     public SyncopeCoreInfoContributor syncopeCoreInfoContributor() {
-        return new SyncopeCoreInfoContributor();
+        return new DefaultSyncopeCoreInfoContributor();
     }
 
     @ConditionalOnMissingBean
@@ -104,5 +109,32 @@ public class SyncopeCoreApplication extends SpringBootServletInitializer {
     @Bean
     public KeymasterStop keymasterStop() {
         return new KeymasterStop(NetworkService.Type.CORE);
+    }
+
+
+    @Bean
+    public SyncopeStarterEventListener syncopeCoreEventListener(
+        @Qualifier("syncopeCoreInfoContributor")
+        final SyncopeCoreInfoContributor syncopeCoreInfoContributor) {
+        return new DefaultSyncopeStarterEventListener(syncopeCoreInfoContributor);
+    }
+
+    @FunctionalInterface
+    public interface SyncopeStarterEventListener {
+        void addLoadInstant(PayloadApplicationEvent<SystemInfo.LoadInstant> event);
+    }
+
+    public static class DefaultSyncopeStarterEventListener implements SyncopeStarterEventListener {
+        private final SyncopeCoreInfoContributor contributor;
+
+        public DefaultSyncopeStarterEventListener(final SyncopeCoreInfoContributor contributor) {
+            this.contributor = contributor;
+        }
+
+        @EventListener
+        @Override
+        public void addLoadInstant(final PayloadApplicationEvent<SystemInfo.LoadInstant> event) {
+            contributor.addLoadInstant(event);
+        }
     }
 }
